# openapi.yaml
openapi: 3.0.3
info:
  title: Notes API
  description: API for managing notes
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /:
    get:
      summary: Test API
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: API works
  /notes/{id}:
    get:
      summary: Get notes by user ID
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteInput'
      responses:
        '201':
          description: Note created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update a note
      parameters:
        - name: id
          in: path
          description: Note ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteInput'
      responses:
        '200':
          description: Note updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a note
      parameters:
        - name: id
          in: path
          description: Note ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Note deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Note:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        content:
          type: string
        uid:
          type: string
    NoteInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        uid:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string